+++ samples/gen-go-db-custom-path/server/router.go
@@ -15,7 +15,7 @@ import (
 	"time"
 
 	"github.com/Clever/go-process-metrics/metrics"
-	"github.com/Clever/wag/samples/v8/gen-go-db-custom-path/tracing"
+	"github.com/Clever/wag/samples/v8/gen-go-db-custom-path/servertracing"
 	"github.com/gorilla/handlers"
 	"github.com/gorilla/mux"
 	"github.com/kardianos/osext"
